+++ AU/Public/Update-AUPackages.ps1
@@ -4,22 +4,30 @@
 <#
 .SYNOPSIS
     Update all automatic packages
+
 .DESCRIPTION
     Function Update-AUPackages will iterate over update.ps1 scripts and execute each. If it detects
-    that package is updated it will push it to Chocolatey repository. The function will look for AU 
-    packages in the current directory or in the directory pointed to by the global variable au_root.
+    that a package is updated it will push it to the Chocolatey community repository.
+
+    The function will look for AU packages in the directory pointed to by the global variable au_root
+    or in the current directory if mentioned variable is not set.
 
     For the push to work, specify your API key in the file 'api_key' in the script's directory or use
-    cached nuget API key or set environment variable '$Env:api_key'
+    cached nuget API key or set environment variable '$Env:api_key'.
+
+    The function accepts many options via ordered HashTable parameter Options.
+
 .EXAMPLE
     Update-AUPackages p* @{ Threads = 5; Timeout = 10 }
 
     Update all automatic packages in the current directory that start with letter 'p' using 5 threads
     and web timeout of 10 seconds.
+
 .EXAMPLE
     $au_root = 'c:\chocolatey'; updateall @{ Force = $true }
 
     Force update of all automatic ackages in the given directory.
+
 .LINK
     Update-Package
 
@@ -36,15 +44,21 @@ function Update-AUPackages {
         Hashtable with options:
           Threads           - Number of background jobs to use, by default 10.
           Timeout           - WebRequest timeout in seconds, by default 100.
-          UpdateTimeout     - Timeout for thread in seconds, by default 1200 (20 minutes).
+          UpdateTimeout     - Timeout for background job in seconds, by default 1200 (20 minutes).
           Force             - Force package update even if no new version is found.
-          Push              - Set to true to push updated packages to chocolatey repository.
-          Mail              - Hashtable with mail notification options: To, Server, UserName, Password, Port, EnableSsl.
+          Push              - Set to true to push updated packages to Chocolatey community repository.
           PluginPath        - Additional path to look for user plugins. If not set only module integrated plugins will work
+
           Plugin            - Any HashTable key will be treated as plugin with the same name as the option name.
-                              Script with that name will be looked into AU module path and user specified path and if
-                              found, it will be called with splatted HashTable passed as parameters.
-          BeforeEach        - User ScriptBlock that will be called before each pacakge, accepts 2 arguments: Name & Options
+                              A script with that name will be searched for in the AU module path and user specified path.
+                              If script is found, it will be called with splatted HashTable passed as plugin parameters.
+
+                              To list default AU plugins run:
+
+                                    ls "$(Split-Path (gmo au -list).Path)\Plugins\*.ps1"
+
+          BeforeEach        - User ScriptBlock that will be called before each package and accepts 2 arguments: Name & Options.
+                              To pass additional arguments, specify them as Options key/values.
           AfterEach         - Similar as above.
           Script            - Script that will be called before and after everything.
         #>